requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.1
packaging==24.0
python-dotenv==1.0.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
wheel==0.43.0
